refactor(christmas): Enum, Record 디렉토리 삭제
