+++ src/main/java/christmas/Process.java
@@ -4,10 +4,10 @@ import christmas.controller.FoodsController;
 import christmas.controller.Validator;
 import christmas.domain.GenerateBenefitStatus;
 import christmas.controller.DayController;
-import christmas.domain.Record.BenefitStatus;
+import christmas.domain.BenefitStatus;
 import christmas.domain.GenerateOrderStatus;
-import christmas.domain.Record.OrderStatus;
-import christmas.domain.Record.WootecoMenu;
+import christmas.domain.OrderStatus;
+import christmas.domain.WootecoMenu;
 import christmas.view.InputView;
 import christmas.view.OutputView;
 import java.util.List;
