+++ src/main/java/christmas/domain/BenefitStatus.java
@@ -1,4 +1,4 @@
-package christmas.domain.Record;
+package christmas.domain;
 
 import java.util.Map;
