+++ src/test/java/christmas/BenefitTest.java
@@ -6,9 +6,9 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import christmas.domain.GenerateBenefitStatus;
 import christmas.domain.GenerateOrderStatus;
 import christmas.domain.GenerateWootecoMenu;
-import christmas.domain.Record.BenefitStatus;
-import christmas.domain.Record.OrderStatus;
-import christmas.domain.Record.WootecoMenu;
+import christmas.domain.BenefitStatus;
+import christmas.domain.OrderStatus;
+import christmas.domain.WootecoMenu;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
